Move <CommentsNavigation> up to <Comments>.
