+++ client/my-sites/comments/comments-navigation/comments-navigation-tab.jsx
@@ -7,10 +7,14 @@
 import React from 'react';
 import classNames from 'classnames';
 
-export const CommentNavigationTab = ( { children, className, onClick } ) => (
+/**
+ * Internal dependencies
+ */
+
+export const CommentsNavigationTab = ( { children, className, onClick } ) => (
 	<div className={ classNames( 'comment-navigation__tab', className ) } onClick={ onClick }>
 		{ children }
 	</div>
 );
 
-export default CommentNavigationTab;
+export default CommentsNavigationTab;
