+++ client/my-sites/comments/style.scss
@@ -1,5 +1,6 @@
-.comment-navigation {
-	@include breakpoint( "<480px" ) {
+/** @format */
+.comments-navigation {
+	@include breakpoint( '<480px' ) {
 		.section-nav__mobile-header:after {
 			padding-left: 8px;
 		}
@@ -8,7 +9,7 @@
 			width: auto;
 		}
 
-		.comment-navigation__open-bulk {
+		.comments-navigation__open-bulk {
 			padding-top: 9px;
 			position: absolute;
 			right: 0;
@@ -17,19 +18,19 @@
 	}
 }
 
-.comment-navigation.is-bulk-edit {
+.comments-navigation.is-bulk-edit {
 	padding-right: 0;
 
-	@include breakpoint( "<480px" ) {
+	@include breakpoint( '<480px' ) {
 		.section-nav__mobile-header {
 			display: none;
 		}
 
-		.comment-navigation__bulk-count {
-			width: calc( 100% - 70px );
+		.comments-navigation__bulk-count {
+			width: calc(100% - 70px);
 		}
 
-		.comment-navigation__close-bulk {
+		.comments-navigation__close-bulk {
 			position: absolute;
 			right: 0;
 			top: 3px;
@@ -38,19 +39,19 @@
 	}
 }
 
-.comment-navigation__bulk-count {
+.comments-navigation__bulk-count {
 	display: inline-flex;
 }
 
-.comment-navigation__bulk-count .count {
+.comments-navigation__bulk-count .count {
 	margin-left: 8px;
 }
 
-.comment-navigation__tab {
+.comments-navigation__tab {
 	padding: 16px;
 }
 
-.comment-navigation__actions {
+.comments-navigation__actions {
 	display: flex;
 	padding: 11px 16px;
 
@@ -59,11 +60,11 @@
 	}
 }
 
-.comment-navigation__open-bulk .button {
+.comments-navigation__open-bulk .button {
 	margin-left: 16px;
 }
 
-.comment-navigation__close-bulk {
+.comments-navigation__close-bulk {
 	margin-left: auto;
 	padding: 0;
 	a {
@@ -81,7 +82,7 @@
 	display: none;
 
 	&.comment-list__transition-enter-active {
-		animation: comment-list__transition-enter .15s linear;
+		animation: comment-list__transition-enter 0.15s linear;
 	}
 }
 
@@ -90,7 +91,7 @@
 
 	&.comment-list__transition-leave-active {
 		opacity: 0.01;
-		transition: opacity .15s linear;
+		transition: opacity 0.15s linear;
 	}
 }
